Refactor StreamingCNN engine modules
