+++ lightstream/core/engine/statistics.py
@@ -1,0 +1,409 @@
+"""Statistics hook registration and collection for the streaming engine."""
+import logging
+from typing import List
+
+import torch
+import torch.nn.functional
+
+from lightstream.core.scnn.utils import Box, Lost, _ntuple, _new_value_indices, H_DIM, W_DIM
+from lightstream.core.scnn.streamingconv import StreamingConv2d
+from lightstream.core.scnn.streamingupsample import StreamingUpsample2d
+from lightstream.core.reducer import BaseReducer
+
+logger = logging.getLogger(__name__)
+
+_triple = _ntuple(3)
+
+
+class StatisticsMixin:
+    def _gather_backward_statistics(self, tile):
+        # Forward pass with grads enabled
+        torch.set_grad_enabled(True)
+        output = self.stream_module(tile)
+        output_tensors, output_spec = self._flatten_output_structure(output)
+        self._output_spec = output_spec
+
+        # Gather backward statistics
+        self._tile_output_shapes = [out.shape for out in output_tensors]
+        self._tile_output_shape = self._tile_output_shapes[0]
+        self._output_stride_per_output = []
+        gradients = []
+        for idx, out in enumerate(output_tensors):
+            lost = self._tile_output_lost[idx]
+            gradient = torch.zeros(*out.shape, dtype=self.dtype, device=self.device)
+            gradient[
+                :,
+                :,
+                lost.top : out.shape[H_DIM] - lost.bottom,
+                lost.left : out.shape[W_DIM] - lost.right,
+            ] = 1
+            gradients.append(gradient)
+
+            p_stats = self._prev_stats(out)
+            if p_stats:
+                output_stride = p_stats["output_stride"] * torch.tensor(p_stats["stride"])
+            else:
+                output_stride = torch.tensor([1, 1, 1])
+
+            self._output_stride_per_output.append(output_stride)
+
+        self.output_stride = self._output_stride_per_output[0]
+        self._base_output_stride = self._output_stride_per_output[0].clone()
+        for stride in self._output_stride_per_output[1:]:
+            self._base_output_stride[1] = min(int(self._base_output_stride[1]), int(stride[1]))
+            self._base_output_stride[2] = min(int(self._base_output_stride[2]), int(stride[2]))
+        torch.autograd.backward(output_tensors, gradients)
+
+        # tiles can have -1, see backward_statistics_hook
+        self.tile_gradient_lost = self._non_max_border_amount(tile.grad)
+
+        # lost statistics assume you're always in the middle of an image, so left,bottom,top,right lost can always happen
+        self._print_verbose("\n", "Input gradient lost", self.tile_gradient_lost)
+
+    def _gather_forward_statistics(self, tile):
+        torch.set_grad_enabled(False)
+        output = self.stream_module(tile)
+        output_tensors, output_spec = self._flatten_output_structure(output)
+        self._output_spec = output_spec
+        self._tile_output_lost = [self._non_max_border_amount(out) for out in output_tensors]
+        self.tile_output_lost = self._tile_output_lost[0]
+        self._print_verbose("\n", "Output lost", self._tile_output_lost)
+
+    def _non_max_border_amount(self, tensor):
+        # Sum over the channels, useful for networks that treat certain channels
+        # different (e.g., DenseNet)
+        if tensor.dim() > 3:
+            tensor = torch.sum(tensor, dim=1)[0]
+        tensor = tensor / tensor.max()  # normalize
+        tensor = tensor > tensor.max() * (1 - self.eps)
+        non_zero = tensor.nonzero(as_tuple=False)
+        top, left = non_zero.min(dim=0)[0]
+        # for bottom and right we need to substract -1: correct index 3 is actually the 4th pixel
+        bottom, right = (
+            torch.tensor([*tensor.size()], dtype=torch.long, device=self.device) - non_zero.max(dim=0)[0] - 1
+        )
+        return Lost(int(top), int(left), int(bottom), int(right))
+
+    def _add_hooks_for_statistics(self):
+        def forw_lambda(module, inpt, outpt):
+            self._forward_gather_statistics_hook(module, inpt, outpt)
+
+        def back_lambda(module, grad_in, grad_out):
+            return self._backward_gather_statistics_hook(module, grad_in, grad_out)
+
+        self._add_hooks(forward_hook=forw_lambda, backward_hook=back_lambda)
+
+    def _add_hooks_for_streaming(self):
+        if self.gather_input_gradient:
+
+            def back_lambda(module, grad_in, grad_out):
+                return self._backward_saliency_hook(module, grad_in, grad_out)
+
+            for mod in self.stream_module.modules():
+                if isinstance(mod, StreamingConv2d):
+                    if mod.in_channels == 3:
+                        self.saliency_input_module = mod
+                        back_handle = mod.register_full_backward_hook(back_lambda)
+                        self._hooks.append(back_handle)
+
+    def _add_hooks(
+        self,
+        forward_hook,
+        backward_hook,
+        forward_modules=(torch.nn.Conv2d, torch.nn.MaxPool2d, torch.nn.AvgPool2d, torch.nn.Upsample),
+        back_modules=(torch.nn.Conv2d, torch.nn.MaxPool2d, torch.nn.Upsample),
+    ):
+        for mod in self.stream_module.modules():
+            if isinstance(mod, forward_modules):
+                forw_handle = mod.register_forward_hook(forward_hook)
+                self._hooks.append(forw_handle)
+                if back_modules and isinstance(mod, back_modules):
+                    back_handle = mod.register_full_backward_hook(backward_hook)
+                    self._hooks.append(back_handle)
+
+    def _remove_hooks(self):
+        for hook in self._hooks:
+            hook.remove()
+
+    def _resolve_upsample_scale(self, module, inpt, output):
+        if module.scale_factor is not None:
+            sf = module.scale_factor
+            if isinstance(sf, tuple):
+                return float(sf[-2]), float(sf[-1])
+            return float(sf), float(sf)
+        in_h, in_w = inpt[0].shape[H_DIM], inpt[0].shape[W_DIM]
+        out_h, out_w = output.shape[H_DIM], output.shape[W_DIM]
+        return float(out_h) / float(max(1, in_h)), float(out_w) / float(max(1, in_w))
+
+    def _update_output_stride_for_upsample(self, prev_output_stride, scale_h, scale_w):
+        out_stride = prev_output_stride.clone().detach().to(torch.float32)
+        out_stride[1] = max(1.0, out_stride[1] / max(scale_h, 1e-8))
+        out_stride[2] = max(1.0, out_stride[2] / max(scale_w, 1e-8))
+        out_stride = torch.round(out_stride).to(torch.long)
+        out_stride[0] = 1
+        return out_stride
+
+    def _forward_gather_statistics_hook(self, module, inpt, output):
+        is_upsample = isinstance(module, torch.nn.Upsample)
+        if not is_upsample:
+            stride, kernel_size, _ = (_triple(module.stride), _triple(module.kernel_size), _triple(module.padding))
+        else:
+            stride = torch.tensor([1, 1, 1])
+            kernel_size = torch.tensor([1, 1, 1])
+
+        if not torch.is_grad_enabled():  # type:ignore
+            # Convert strided convolutions/pooling to average pool
+            if (not is_upsample) and (
+                isinstance(module, (torch.nn.MaxPool2d))
+                or (stride[0] > 1 and stride[0] > kernel_size[0])
+                or (stride[1] > 1 and stride[1] > kernel_size[1])
+                or (stride[2] > 1 and stride[2] > kernel_size[2])
+            ):
+                # Pytorch documentation is explicitely against changing output in a forward hook
+                # However, since we do not really need the graph or gradients to be correct
+                # it shouldn't harm.
+                if module.padding != 0:
+                    padding = module.padding
+                    if not isinstance(module.padding, tuple):
+                        padding = [module.padding, module.padding]
+                    padded_input = torch.nn.functional.pad(inpt[0], [padding[1], padding[1], padding[0], padding[0]])
+                else:
+                    padded_input = inpt[0]
+
+                new_output = torch.nn.functional.avg_pool2d(padded_input, kernel_size[1:], stride[1:])
+                new_output = torch.sum(new_output, dim=1)[0]
+                new_output = new_output > (1 - self.eps) * new_output.max()
+                new_output = new_output.expand_as(output[0])
+
+                output[0] = new_output.type(self.dtype)
+
+            # Sum all dimensions (useful for DenseNet like networks)
+            lost = self._non_max_border_amount(output)
+
+            # Make output between 0-1 again, so the values do not explode
+            output.fill_(0)
+            output[
+                :, :, lost.top : output[0, 0].shape[0] - lost.bottom, lost.left : output[0, 0].shape[1] - lost.right
+            ] = 1
+
+            module_stats = {"lost": lost, "stride": stride if not is_upsample else torch.tensor([1, 1, 1]), "module": module}
+            self._print_verbose(module, "\n", module_stats["lost"])
+
+            self._saved_tensors[module] = inpt
+            self._module_stats[module] = module_stats
+        else:
+            module_stats = self._module_stats[module]
+
+            p_stats = self._prev_stats(output)
+            if p_stats:
+                prev_output_stride = p_stats["output_stride"] * p_stats["stride"].clone().detach() if isinstance(p_stats["stride"], torch.Tensor) else p_stats["output_stride"] * torch.tensor(p_stats["stride"])
+            else:
+                prev_output_stride = torch.tensor([1, 1, 1])
+
+            if is_upsample:
+                scale_h, scale_w = self._resolve_upsample_scale(module, inpt, output)
+                output_stride = self._update_output_stride_for_upsample(prev_output_stride, scale_h, scale_w)
+                module_stats["scale_factor_hw"] = (scale_h, scale_w)
+            else:
+                output_stride = prev_output_stride
+
+            module_stats["output_stride"] = output_stride.clone().detach()
+            self._stats_per_grad_fn[output.grad_fn] = module_stats
+            self._module_stats[module] = module_stats
+
+    def _backward_gather_statistics_hook(self, module, grad_in, grad_out):
+        is_upsample = isinstance(module, torch.nn.Upsample)
+        if not is_upsample:
+            stride, kernel_size, _ = (_triple(module.stride), _triple(module.kernel_size), _triple(module.padding))
+        if grad_in[0] is not None:
+            # We sum over the channels to deal with networks that do different operations
+            # on groups of channels
+            f_grad = torch.sum(grad_in[0], dim=1)[0]
+            if isinstance(module, (torch.nn.MaxPool2d)):
+                # MaxPool shifts indices around, which break the calculation to
+                # find valid gradient values. To fix this we do an average pool
+                # with the same kernel-size and stride and repeat using the stride.
+
+                # We have the input gradient for a max pool right now
+                # The following computes the input gradient for an average pool instead
+                # compute the output of the average pool, and multiply by output gradient
+
+                # from forward statistics hook, modified input
+                inpt = self._saved_tensors[module]
+                padded_inpt = inpt[0]
+
+                if module.padding != 0:
+                    padded_inpt = torch.nn.functional.pad(
+                        inpt[0], [module.padding, module.padding, module.padding, module.padding], value=-1
+                    )
+
+                new_outpt = torch.nn.functional.avg_pool2d(padded_inpt, kernel_size[1:], stride[1:])[0]
+                new_outpt = torch.sum(new_outpt, dim=0)
+
+                f_grad = torch.sum(grad_out[0], dim=1)[0]
+                f_grad = f_grad * new_outpt
+                f_grad = f_grad.repeat_interleave(stride[1], dim=0)
+                f_grad = f_grad.repeat_interleave(stride[2], dim=1)
+                grad = torch.zeros(grad_in[0].shape[2:], dtype=f_grad.dtype, device=f_grad.device)
+
+                self._print_verbose("testing shape gradient fix")
+                grad[: f_grad.shape[0], : f_grad.shape[1]] = f_grad[: grad.shape[0], : grad.shape[1]]
+
+                f_grad = grad.to(self.device)
+
+            if grad_out[0].numel() == 0 or torch.count_nonzero(grad_out[0]) == 0:
+                # Some connected branches (e.g. zero-scaled passthrough links for graph connectivity)
+                # produce valid but all-zero gradients during stats gathering; skip border inference.
+                return grad_in
+
+            grad_lost = self._non_max_border_amount(grad_out[0])
+
+            self._print_verbose(module, "\n", grad_lost)
+            self._module_stats[module]["grad_lost"] = grad_lost
+
+            valid_grad = f_grad > (1 - self.eps) * f_grad.max()
+
+            # When kernel_size > stride we have some _overlap_ of gradients,
+            # this overlap makes extra positions in the input gradient invalid
+            if (not is_upsample) and (
+                (stride[0] > 1 and kernel_size[0] > stride[0])
+                or (stride[1] > 1 and kernel_size[1] > stride[1])
+                or (stride[2] > 1 and kernel_size[2] > stride[2])
+            ):
+                valid_lost = self._non_max_border_amount(f_grad)
+                valid_grad.fill_(0)
+                overlap_rows = kernel_size[1] - stride[1]
+                overlap_cols = kernel_size[2] - stride[2]
+                valid_grad[
+                    valid_lost.top + overlap_rows : valid_grad.shape[0] - valid_lost.bottom - overlap_rows,
+                    valid_lost.left + overlap_cols : valid_grad.shape[1] - valid_lost.right - overlap_cols,
+                ] = 1
+
+            new_grad_in = valid_grad[None].expand(grad_in[0].shape[1], *valid_grad.shape)[None]
+            new_grad_in = new_grad_in.type(self.dtype) * 10 - 1
+            new_grad_in_lost = self._non_max_border_amount(new_grad_in)
+
+            return (new_grad_in, *grad_in[1:])
+
+    def _backward_saliency_hook(self, module: StreamingConv2d, grad_in, grad_out, is_bias=False, change_grad=True):
+        stride: List[int] = _triple(module.stride)  # type:ignore
+
+        # Trim gradient of invalid values
+        sides = module.input_loc.sides
+        grad_lost = module.grad_lost  # type: Lost
+
+        lost_top = grad_lost.top if not sides.top else 0
+        lost_bottom = grad_lost.bottom if not sides.bottom else 0
+        lost_left = grad_lost.left if not sides.left else 0
+        lost_right = grad_lost.right if not sides.right else 0
+        lost = Lost(lost_top, lost_left, lost_bottom, lost_right)
+
+        grad = grad_out[0]
+        valid_grad = grad[:, :, lost_top : grad.shape[H_DIM] - lost_bottom, lost_left : grad.shape[W_DIM] - lost_right]
+
+        output_stride = module.output_stride * torch.tensor(stride)
+        input_loc = module.input_loc
+
+        # Move the location according to how many pixels have been trimmed
+        # this will be the location of the valid gradient of this layer in relation
+        # to the actual gradient in a normal backpass
+        data_loc_y = int(input_loc.y // output_stride[1]) + lost_top
+        data_loc_x = int(input_loc.x // output_stride[2]) + lost_left
+
+        data_loc = Box(data_loc_y, 0, data_loc_x, 0, input_loc.sides)
+
+        # Calculate which part of the gradient is 'new'
+        old_value_indices = self.saliency_old_indices
+        new_output_box, updated_total_indices = _new_value_indices(valid_grad.shape, data_loc, old_value_indices)
+
+        if module.in_channels == 3:
+            valid_grad_in = grad_in[0][
+                :,
+                :,
+                lost.top * stride[1] : grad_in[0].shape[2] - lost.bottom * stride[1],
+                lost.left * stride[2] : grad_in[0].shape[3] - lost.right * stride[2],
+            ]
+
+            relevant_input_grad = valid_grad_in[
+                :,
+                :,
+                new_output_box.y * stride[1] : new_output_box.y * stride[1] + new_output_box.height * stride[1],
+                new_output_box.x * stride[2] : new_output_box.x * stride[2] + new_output_box.width * stride[2],
+            ]
+
+            self.saliency_map[
+                :,
+                :,
+                updated_total_indices.y * stride[1] : updated_total_indices.height * stride[1],
+                updated_total_indices.x * stride[2]
+                - relevant_input_grad.shape[3] : updated_total_indices.x * stride[2],
+            ] = relevant_input_grad.detach().cpu()
+
+            del relevant_input_grad
+            del valid_grad_in
+        return grad_in
+
+    def _prev_stats(self, grad_fn):
+        """DAG traversal, finds the first grad_fn that is in self._stats_per_grad_fn
+
+        Finds the first grad_fn that is in self._stats_per_grad_fn, which is needed for output stride calculations
+
+        Parameters
+        ----------
+        grad_fn: the grad function of the current output tensor
+
+        """
+        if hasattr(grad_fn, "grad_fn"):
+            grad_fn = grad_fn.grad_fn
+
+        prev_stats = None
+
+        if grad_fn in self._stats_per_grad_fn:
+            prev_stats = self._stats_per_grad_fn[grad_fn]
+            return prev_stats
+        elif hasattr(grad_fn, "next_functions") and len(grad_fn.next_functions) > 0:
+            children = [x[0] for x in grad_fn.next_functions]
+
+            for x in children:
+                prev_stats = self._prev_stats(x)
+                if prev_stats is not None:
+                    break
+            return prev_stats
+        return prev_stats
+
+    def get_tile_cache(self):
+        named_stats = {"net_stats": {}}
+        for name, module in self.stream_module.named_modules():
+            if module in self._module_stats:
+                named_stats["net_stats"][name] = self._module_stats[module]
+        named_stats["output_stride"] = self.output_stride
+        named_stats["tile_output_lost"] = self.tile_output_lost  # type:ignore
+        named_stats["tile_output_lost_all"] = self._tile_output_lost  # type:ignore
+        named_stats["tile_gradient_lost"] = self.tile_gradient_lost  # type:ignore
+        named_stats["tile_output_shape"] = self._tile_output_shape  # type:ignore
+        named_stats["tile_output_shapes"] = self._tile_output_shapes  # type:ignore
+        named_stats["output_stride_per_output"] = self._output_stride_per_output  # type:ignore
+        named_stats["output_spec"] = self._output_spec
+        return named_stats
+
+    def load_tile_cache(self, state):
+        self.disable()
+
+        self.output_stride = state["output_stride"]
+        self.tile_output_lost = state["tile_output_lost"]
+        self._tile_output_lost = state.get("tile_output_lost_all", [self.tile_output_lost])
+        self.tile_gradient_lost = state["tile_gradient_lost"]
+        self._tile_output_shape = state["tile_output_shape"]
+        self._tile_output_shapes = state.get("tile_output_shapes", [self._tile_output_shape])
+        self._output_stride_per_output = state.get("output_stride_per_output", [self.output_stride])
+        self._base_output_stride = self._output_stride_per_output[0].clone()
+        for stride in self._output_stride_per_output[1:]:
+            self._base_output_stride[1] = min(int(self._base_output_stride[1]), int(stride[1]))
+            self._base_output_stride[2] = min(int(self._base_output_stride[2]), int(stride[2]))
+        self._output_spec = state.get("output_spec", ("tensor", None))
+
+        for name, module in self.stream_module.named_modules():
+            if name in state["net_stats"]:
+                self._module_stats[module] = state["net_stats"][name]
+
+        self.enable()
